// src/app/models/InsContract.ts
export class InsContract {
    idCont?: number;
    insIdCont?: number;
    idEapb?: number;
    idTama?: number;
    nameCont?: string;    
    intialDateCont?: Date;
    finalDateCont?: Date;
    regimeCont?: number;
    kindCont?: number;
    codeCont?: string;
    stateCont?:string;
    totalValueContract?:number;
    ajustTama?:number;

  }